useEnterpriseData for fetch
